Day 9
Day 9

+++ Day_8_Task_4/main.c
@@ -3,7 +3,7 @@
 
 int main()
 {
-    long int a;
+    int a;
     char c;
     float b;
     double f;
